fix: Change ff and improper periodicity assertion statement. fixes  #27 (#30)
* Change ff to amber99sb-ildn

Fixes partially #28\n\nThe ff amber99sb-star-ildnp is not per default shipped with gromacs as dissussed in #24

* Fix assertion statement regarding the number of periodicity terms for impropers

The number of periodicity terms for proper/improper dihedrals is a model hyperparameter and may not be equal to 3 for impropers. The number of periodicity terms is now determined from the model used.